requirements.txt
requests
pydantic
python-dotenv
langchain-google-genai 
langchain-openai 
crewai-tools